渲染引擎

+++ Substance/Maths/Vector2T.cs
@@ -1,5 +1,5 @@
 global using Vector2Int = Substance.Maths.Vector2T<int>;
-global using Vector2 = Substance.Maths.Vector2T<int>;
+global using Vector2 = Substance.Maths.Vector2T<float>;
 
 using System.Numerics;
